+++ src/graphix_lab/domain/__init__.py
@@ -1,1 +1,19 @@
 """Domain models for the template package."""
+
+from .errors import (
+    GraphixCompatibilityError,
+    GraphixLabError,
+    GraphixUnavailableError,
+    OptionalDependencyError,
+    UnsupportedBackendError,
+    UnsupportedGateError,
+)
+
+__all__ = [
+    "GraphixCompatibilityError",
+    "GraphixLabError",
+    "GraphixUnavailableError",
+    "OptionalDependencyError",
+    "UnsupportedBackendError",
+    "UnsupportedGateError",
+]